requirements.txt
tensorflow-cpu==2.16.1
pandas
numpy<2.0
scikit-learn
matplotlib
seaborn
streamlit
tensorboard
ipykernel
